+++ lib/maze/client/appium/bs_client.rb
@@ -36,6 +36,10 @@ module Maze
 
         def log_session_info
           if Maze.config.device || Maze.config.browser
+            # Log device id
+            udid = Maze.driver.session_capabilities['udid']
+            $logger.info "Running on device: #{udid}" unless udid.nil?
+
             # Log a link to the BrowserStack session search dashboard
             build = Maze.driver.capabilities[:build]
             url = "https://app-automate.browserstack.com/dashboard/v2/search?query=#{build}&type=builds"